+++ tubular/mixins.py
@@ -110,17 +110,6 @@ class DropOriginalMixin:
         return X.to_native() if return_native else X
 
 
-class NewColumnNameMixin:
-    """Helper to validate and set new_column_name attribute"""
-
-    def check_and_set_new_column_name(self, new_column_name: str) -> None:
-        if not (isinstance(new_column_name, str)):
-            msg = f"{self.classname()}: new_column_name should be str"
-            raise TypeError(msg)
-
-        self.new_column_name = new_column_name
-
-
 class SeparatorColumnMixin:
     """Hel per to validate and set separator attribute"""
 
@@ -130,19 +119,6 @@ class SeparatorColumnMixin:
             raise TypeError(msg)
 
         self.separator = separator
-
-
-class TwoColumnMixin:
-    """helper to validate columns when exactly two columns are required"""
-
-    def check_two_columns(self, columns: list[str]) -> None:
-        if not (isinstance(columns, list)):
-            msg = f"{self.classname()}: columns should be list"
-            raise TypeError(msg)
-
-        if len(columns) != 2:
-            msg = f"{self.classname()}: This transformer works with two columns only"
-            raise ValueError(msg)
 
 
 class WeightColumnMixin: